Added a new SQL view and increased size of Version column in Invocations Store

+++ src/Services/Work/NuGet.Services.Work.Database/private.table.InvocationsStore.sql
@@ -3,7 +3,7 @@
 -- Use SPROCS!
 CREATE TABLE [private].[InvocationsStore]
 (
-    [Version] int NOT NULL PRIMARY KEY IDENTITY,
+    [Version] BIGINT NOT NULL PRIMARY KEY IDENTITY,
     [Id] uniqueidentifier NOT NULL,
     [Job] nvarchar(50) NOT NULL,
     [Source] nvarchar(50) NOT NULL,
